requirements.txt
opencv-python
ultralytics
paho-mqtt
python-dotenv
boto3 
pytesseract
tesseract